Chapter 2, first draft
Intial commit of chapter 2, which deals with b2, Michel Valdrighi, and
b2 becoming dormant. Note that as yet, no interview with Michel has
been carried out (beyond a brief Skype chat. Also, the b2 forums are
still missing and it looks unlikely that we will get them. The majority
of the information in this chapter comes from Michel's blog, other blog
posts, and a presentation that Michel did at WordCamp Paris.